(PC-7477) Move useItinerary under the libs/ folder
